requirements.txt
Django==4.2.6
djangorestframework==3.14.0
PyYAML==6.0.1
psycopg2==2.9.9
pandas==2.1.2
django-cors-headers==4.3.0
redis==5.0.1
whitenoise==6.6.0
boto3==1.22.8
django-storages==1.13.2
celery[sqs]==5.2.6
beautifulsoup4==4.12.3
lxml==5.1.0
PyJWT==2.8.0
requests==2.31.0